[Enhancement] Move dictionary
