+++ mmocr/models/common/__init__.py
@@ -1,5 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .backbones import *  # NOQA
+from .dictionary import *  # NOQA
 from .layers import *  # NOQA
 from .losses import *  # NOQA
 from .modules import *  # NOQA
