+++ mmocr/models/kie/heads/sdmgr_head.py
@@ -6,7 +6,7 @@ from mmengine.model import BaseModule
 from torch import Tensor, nn
 from torch.nn import functional as F
 
-from mmocr.models.textrecog.dictionary import Dictionary
+from mmocr.models.common.dictionary import Dictionary
 from mmocr.registry import MODELS, TASK_UTILS
 from mmocr.structures import KIEDataSample
