+++ mmocr/models/textrecog/__init__.py
@@ -2,7 +2,6 @@
 from .backbones import *  # NOQA
 from .data_preprocessors import *  # NOQA
 from .decoders import *  # NOQA
-from .dictionary import *  # NOQA
 from .encoders import *  # NOQA
 from .module_losses import *  # NOQA
 from .plugins import *  # NOQA
