+++ mmocr/models/textrecog/decoders/abi_vision_decoder.py
@@ -5,8 +5,8 @@ import torch
 import torch.nn as nn
 from mmcv.cnn import ConvModule
 
+from mmocr.models.common.dictionary import Dictionary
 from mmocr.models.common.modules import PositionalEncoding
-from mmocr.models.textrecog.dictionary import Dictionary
 from mmocr.registry import MODELS
 from mmocr.structures import TextRecogDataSample
 from .base_decoder import BaseDecoder
