+++ mmocr/models/textrecog/decoders/base_decoder.py
@@ -4,7 +4,7 @@ from typing import Dict, List, Optional, Sequence, Union
 import torch
 from mmengine.model import BaseModule
 
-from mmocr.models.textrecog.dictionary import Dictionary
+from mmocr.models.common.dictionary import Dictionary
 from mmocr.registry import MODELS, TASK_UTILS
 from mmocr.structures import TextRecogDataSample
