+++ mmocr/models/textrecog/decoders/crnn_decoder.py
@@ -5,7 +5,7 @@ import torch
 import torch.nn as nn
 from mmengine.model import Sequential
 
-from mmocr.models.textrecog.dictionary import Dictionary
+from mmocr.models.common.dictionary import Dictionary
 from mmocr.models.textrecog.layers import BidirectionalLSTM
 from mmocr.registry import MODELS
 from mmocr.structures import TextRecogDataSample
