+++ mmocr/models/textrecog/decoders/master_decoder.py
@@ -8,8 +8,8 @@ import torch.nn as nn
 from mmcv.cnn.bricks.transformer import BaseTransformerLayer
 from mmengine.model import ModuleList
 
+from mmocr.models.common.dictionary import Dictionary
 from mmocr.models.common.modules import PositionalEncoding
-from mmocr.models.textrecog.dictionary import Dictionary
 from mmocr.registry import MODELS
 from mmocr.structures import TextRecogDataSample
 from .base_decoder import BaseDecoder
