+++ mmocr/models/textrecog/decoders/robust_scanner_fuser.py
@@ -5,7 +5,7 @@ from typing import Dict, List, Optional, Sequence, Union
 import torch
 import torch.nn as nn
 
-from mmocr.models.textrecog.dictionary import Dictionary
+from mmocr.models.common.dictionary import Dictionary
 from mmocr.registry import MODELS
 from mmocr.structures import TextRecogDataSample
 from .base_decoder import BaseDecoder
