+++ mmocr/models/textrecog/decoders/sequence_attention_decoder.py
@@ -5,7 +5,7 @@ from typing import Dict, Optional, Sequence, Union
 import torch
 import torch.nn as nn
 
-from mmocr.models.textrecog.dictionary import Dictionary
+from mmocr.models.common.dictionary import Dictionary
 from mmocr.models.textrecog.layers import DotProductAttentionLayer
 from mmocr.registry import MODELS
 from mmocr.structures import TextRecogDataSample
