+++ mmocr/models/textrecog/module_losses/ce_module_loss.py
@@ -5,7 +5,7 @@ from typing import Dict, Sequence, Union
 import torch
 import torch.nn as nn
 
-from mmocr.models.textrecog.dictionary.dictionary import Dictionary
+from mmocr.models.common.dictionary import Dictionary
 from mmocr.registry import MODELS
 from mmocr.structures import TextRecogDataSample
 from .base_recog_module_loss import BaseRecogModuleLoss
