+++ mmocr/models/textrecog/postprocessors/base_textrecog_postprocessor.py
@@ -6,7 +6,7 @@ import mmengine
 import torch
 from mmengine.data import LabelData
 
-from mmocr.models.textrecog.dictionary import Dictionary
+from mmocr.models.common.dictionary import Dictionary
 from mmocr.registry import TASK_UTILS
 from mmocr.structures import TextRecogDataSample
