+++ tests/test_models/test_textrecog/test_module_losses/test_ctc_module_loss.py
@@ -6,7 +6,7 @@ from unittest import TestCase
 import torch
 from mmengine.data import LabelData
 
-from mmocr.models.textrecog.dictionary import Dictionary
+from mmocr.models.common.dictionary import Dictionary
 from mmocr.models.textrecog.module_losses import CTCModuleLoss
 from mmocr.structures import TextRecogDataSample
 from mmocr.testing import create_dummy_dict_file
