+++ tests/test_models/test_textrecog/test_postprocessors/test_ctc_postprocessor.py
@@ -6,7 +6,7 @@ from unittest import TestCase
 
 import torch
 
-from mmocr.models.textrecog.dictionary import Dictionary
+from mmocr.models.common.dictionary import Dictionary
 from mmocr.models.textrecog.postprocessors.ctc_postprocessor import \
     CTCPostProcessor
 from mmocr.structures import TextRecogDataSample
